+++ src/tests/transformer_new.cpp
@@ -1,0 +1,11 @@
+#include "marian.h"
+#include "models/transformer_new.h"
+
+
+int main(int argc, char** argv) {
+  using namespace marian;
+
+  testme();
+
+  return 0;
+}